-Bug fix category controller + custom fields. -Aggiunta creazione utenti con possibilità di mantenere solo informazioni su personale e quindi creazione di account non autenticabili. -Predisposizione alla visualizzazione dei cusotm fields nell'editor del post. -Modifica di tutte le migrations con la funzine id() a laravel 8 -Aggiornamento attributi delle colonne al db di alcune tabelle tramite migrations -Aggiunta delle colonne al db tramite migrations su users -Aggiornamento seeds per supportare le modifiche alle colonne delle tabelle -Aggiunta traduzione n a n tra users e category che contiene i valori dei custom fields -Modifica della view di editing della category con controllo su modifica custom fields (se la categoria è associata almeno ad un post, i fields non sono modificabili) -Aggiunta parametri alla localizzazione inglese -Creazione delle view di creazione degli utenti -Aggiornamento di Admin/UserController per supportare la creazione degli utenti
TODO:
-Aggiornamento sistema di login con verifica se l'utente è autenticabile
-Aggiornamento editor dei custom fields con caricamento del valore di
questi
-Modificare le funzionalità dei custom fields dal model post al model
cusotmFields

BUGS:
-In fase di creazione dell'utente, il sistema ha problemi nella gestione
dell'id nella Admin/Requests/UsersRequest
-Varie ed eventuali

+++ database/migrations/2017_03_19_102521_add_thumbnail_id_to_posts_table.php
@@ -14,7 +14,7 @@ class AddThumbnailIdToPostsTable extends Migration
     public function up()
     {
         Schema::table('posts', function (Blueprint $table) {
-            $table->integer('thumbnail_id')->unsigned()->nullable();
+            $table->bigInteger('thumbnail_id')->unsigned()->nullable();
             $table->foreign('thumbnail_id')->references('id')->on('media');
         });
     }
